refactor: entity컬럼 수정 및 타입 변경
- 정확하게 금액을 계산해야하는경우가 있고
정수형 데이터타입을 그냥 쓰지는 않으므로
Long -> BigDecimal로 타입변경
- config,util 구분으로 인한 폴더구성 변겅

+++ src/main/java/com/goodsmall/modules/product/domain/Product.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.persistence.*;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -24,7 +25,7 @@ public class Product {
     private String image;
 
     @Column(name="product_price",nullable = false)
-    private long productPrice;
+    private BigDecimal productPrice;
 
     @Column(name="product_open_date")
     @JsonFormat(pattern = "yyyy-MM-dd kk:mm:ss")
